+++ src/basic/App.tsx
@@ -3,7 +3,8 @@ import { CartItem, Coupon, Product } from "../types";
 import { initialCoupons, initialProducts } from "./constants/data";
 import { ProductWithUI } from "./types/product";
 import { Notification } from "./types";
-import Header from "./components/Header";
+import Header from "./components/layout/Header";
+import Toast from "./components/elements/Toast";
 
 const App = () => {
   const [products, setProducts] = useState<ProductWithUI[]>(() => {
@@ -359,23 +360,7 @@ const App = () => {
 
   return (
     <div className="min-h-screen bg-gray-50">
-      {notifications.length > 0 && (
-        <div className="fixed top-20 right-4 z-50 space-y-2 max-w-sm">
-          {notifications.map((notif) => (
-            <div
-              key={notif.id}
-              className={`p-4 rounded-md shadow-md text-white flex justify-between items-center ${notif.type === "error" ? "bg-red-600" : notif.type === "warning" ? "bg-yellow-600" : "bg-green-600"}`}
-            >
-              <span className="mr-2">{notif.message}</span>
-              <button onClick={() => setNotifications((prev) => prev.filter((n) => n.id !== notif.id))} className="text-white hover:text-gray-200">
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                </svg>
-              </button>
-            </div>
-          ))}
-        </div>
-      )}
+      <Toast notifications={notifications} setNotifications={setNotifications} />
       <Header isAdmin={isAdmin} searchTerm={searchTerm} setSearchTerm={setSearchTerm} setIsAdmin={setIsAdmin} cart={cart} totalItemCount={totalItemCount} />
       <main className="max-w-7xl mx-auto px-4 py-8">
         {isAdmin ? (
