+++ src/basic/components/layout/Header.tsx
@@ -1,4 +1,4 @@
-import { CartItem } from "../../types";
+import { CartItem } from "../../../types";
 
 export interface IHeaderProps {
   isAdmin: boolean;
